requirements.txt
ultralytics>=8.0.0
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
numpy>=1.21.0
PyYAML>=6.0
Pillow>=8.3.0
scipy>=1.8.0
tensorboard>=2.8.0
tqdm>=4.64.0